refactoring part 1.
