+++ src/SfxWeb/src/app/Models/eventstore/timelineGenerators.ts
@@ -6,8 +6,6 @@ import padStart from 'lodash/padStart';
 import findIndex from 'lodash/findIndex';
 import { HtmlUtils } from 'src/app/Utils/HtmlUtils';
 import { RepairTask } from 'src/app/Models/DataModels/repairTask';
-import { IConcurrentEventsConfig, IConcurrentEvents, IRCAItem } from 'src/app/modules/event-store/event-store/event-store.component';
-import { Utils } from 'src/app/Utils/Utils';
 
 /*
     NOTES:
